# backend/docs/swagger.yaml
basePath: /api
definitions:
  internal_api.BulkInvitationRequest:
    properties:
      invitations:
        items:
          $ref: '#/definitions/internal_api.InvitationRow'
        type: array
    type: object
  internal_api.BulkInvitationResponse:
    properties:
      job_id:
        type: string
      message:
        type: string
      status:
        type: string
      total_records:
        type: integer
    type: object
...
  internal_api.InvitationJob:
    properties:
      job_id:
        type: string
      status:
        type: string
      total_records:
        type: integer
    type: object
  internal_api.InvitationRetryResponse:
    properties:
      job_id:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  internal_api.InvitationRow:
    properties:
      email:
        type: string
      full_name:
        type: string
    type: object
...
  internal_api.createClassRequest:
    properties:
      gradeId:
        type: string
      name:
        type: string
      streamId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Platform team
  description: Backend REST API for Somotracker educational dashboard
  license:
    name: MIT
  termsOfService: http://somotracker.local/terms
  title: Somotracker API
  version: 1.0.0
paths:
...
  /api/admins/invitations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bulk invitation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.BulkInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.BulkInvitationResponse'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_api.BulkInvitationResponse'
      summary: Bulk invite admins
      tags:
      - Admin Invitations
  /api/admins/invitations/jobs/{job_id}:
    get:
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.InvitationJob'
      summary: Get invitation job status
      tags:
      - Admin Invitations
  /api/admins/invitations/jobs/{job_id}/events:
    get:
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get invitation job events
      tags:
      - Admin Invitations
  /api/admins/invitations/jobs/{job_id}/retry-failed:
    post:
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.InvitationRetryResponse'
      summary: Retry failed invitation jobs
      tags:
      - Admin Invitations
...
  /school/classes:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Comma separated grade labels
        in: query
        name: grade
        type: string
      - description: Comma separated stream names
        in: query
        name: stream
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: List classes
      tags:
      - Classes
    post:
      consumes:
      - application/json
      parameters:
      - description: Class payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.createClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Create class
      tags:
      - Classes
  /school/classes/{id}:
    get:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get class
      tags:
      - Classes
...
swagger: "2.0"